requirements-dev.txt
-r requirements-base.txt
black==19.10b0
django_extensions==3.0.5
factory-boy==2.12.0
flake8==3.8.3
flake8-blind-except==0.1.1
flake8-debugger==3.2.1
flake8-docstrings==1.5.0
flake8-isort==3.0.1
flake8-string-format==0.3.0
flake8-tuple==0.4.1
gitlint==0.13.1
ipython==7.16.1
isort==4.3.21
lxml==4.6.2
pdbpp==0.10.2
pre-commit==2.8.2
psycopg2-binary==2.8.5
pytest==5.4.3
pytest-cov==2.10.1
pytest-django==3.9.0
pytest-env==0.6.2
pytest-factoryboy==2.0.3
pytest-freezegun==0.4.2
pytest-mock==3.3.0
pytest-randomly==3.4.1
requests-mock==1.8.0
snapshottest==0.5.1
xmlschema==1.3.1
